Refactor useToast hook
Refactored useToast hook: divided structure into separate components,
improved typings
