+++ src/components/Toast/index.ts
@@ -1,0 +1,5 @@
+export { default as Toast } from "./Toast"
+export {
+  default as ToastContainer,
+  toastContainerFactory,
+} from "./ToastContainer"
